--- components/CreateCard/index.js
@@ -1,1 +1,0 @@
-export { default } from "./CreateCard.js";
