--- components/FitButtons/index.js
@@ -1,1 +1,0 @@
-export { default } from "./FitButtons.js";
