--- components/FitMap/index.js
@@ -1,1 +1,0 @@
-export { default } from "./FitMap.js";
