--- components/FitNew/index.js
@@ -1,1 +1,0 @@
-export { default } from "./FitNew.js";
