+++ components/HardQuest/HardQuestStyles.js
