+++ components/Home/HomeStyles.js
