+++ components/NewQuest/NewQuestStyles.js
